Updated UI

+++ app/src/main/java/com/example/placeholder/Cart.java
@@ -1,17 +1,15 @@
 package com.example.placeholder;
-import java.text.NumberFormat;
-
 public class Cart {
     private Item [] order;
-    private double totalPrice;
+    private int totalPrice;
     private int totalCount;
 
     public Cart(){
         order = new Item[40];
-        totalPrice= 0.0;
+        totalPrice= 0;
         totalCount=0;
     }
-    public void addToCart(String itemName, double price){
+    public void addToCart(String itemName, int price){
         Item temp= new Item (itemName, price);
         totalPrice+=price;
         order[totalCount]=temp;
